Version 1.4.4-1 - Some improvements to the man pages (thanks to Ville Skyttä). - Fixed a possible segfault in cSkins::Message() (thanks to Udo Richter). - Made the getskyepg.pl script of the 'sky' plugin send a user agent message to   the server, according to the rules at http://bleb.org/tv/data/listings.   If your version of 'wget' doesn't support the -U option to set the user agent,   use the new option -U of getskyepg.pl to have the information added to the URL   as a query string. - The getskyepg.pl script now replaces "&amp;" with "&". - Fixed a possible crash in remux.c on 64-bit machines (thanks to Reinhard Nissl). - Fixed a typo in the change to the "Use small font" setup option in version 1.3.47   in the HISTORY and CONTRIBUTORS file (reported by Andreas Brugger). - Added a missing 'const' to cRecordingInfo::ChannelID() (reported by Andreas   Brugger). This required the APIVERSION to be increased, so plugins will have to   be recompiled. - Now calling cPluginManager::Active() only if VDR is really trying to shut down,   and waiting for 5 minutes before calling it again (thanks to Jörg Wendel for   reporting that cPlugin::Active() was called too often, and to Udo Richter for   some hints on how to improve this). - Replaced 'unsigned long' with 'uint32_t' and 'uint64' with 'uint64_t' to   avoid problems on 64-bit machines.

+++ remux.c
@@ -11,7 +11,7 @@
  * The cRepacker family's code was originally written by Reinhard Nissl <rnissl@gmx.de>,
  * and adapted to the VDR coding style by Klaus.Schmidinger@cadsoft.de.
  *
- * $Id: remux.c 1.56 2006/04/17 12:48:12 kls Exp $
+ * $Id: remux.c 1.57 2006/12/01 14:46:25 kls Exp $
  */
 
 #include "remux.h"
@@ -389,7 +389,7 @@ bool cVideoRepacker::ScanDataForStartCodeFast(const uchar *&Data, const uchar *L
         }
 
   Data = Limit;
-  unsigned long *Scanner = (unsigned long *)(Data - 3);
+  uint32_t *Scanner = (uint32_t *)(Data - 3);
   scanner = ntohl(*Scanner);
   return false;
 }
@@ -615,7 +615,7 @@ bool cVideoRepacker::ScanForEndOfPictureFast(const uchar *&Data, const uchar *Li
         }
 
   Data = Limit + 1;
-  unsigned long *LocalScanner = (unsigned long *)(Data - 4);
+  uint32_t *LocalScanner = (uint32_t *)(Data - 4);
   localScanner = ntohl(*LocalScanner);
   return false;
 }
